requirements.txt
streamlit==1.12.0
openai==1.3.0
python-dotenv==1.0.0
pandas==1.5.3
numpy==1.24.3
mlflow==2.8.1
dvc==3.30.1
prometheus-client==0.17.1
python-gitlab==3.15.0
PyGithub==2.1.1
plotly==5.18.0
scikit-learn==1.3.2 